implemented getBill() and first tests

+++ src/main/java/io/lab/hotel/iFasada.java
@@ -22,7 +22,7 @@ public interface iFasada {
 
 	boolean fireEmployee(String PESEL);
 
-	float getBill(float reservationID);
+	float getBill(int reservationID);
 
 
 	ArrayList<Employee> getEmployeesList(int hotelID);
